[filter] rename internal namespace for global uniqueness
