+++ documentation/Doxyfile
@@ -181,7 +181,7 @@ RECURSIVE              = YES
 EXCLUDE                =
 EXCLUDE_SYMLINKS       = NO
 EXCLUDE_PATTERNS       =
-EXCLUDE_SYMBOLS        = *::internal::*
+EXCLUDE_SYMBOLS        = *::kalman_internal::*
 EXAMPLE_PATH           = ../sample
 EXAMPLE_PATTERNS       = *
 EXAMPLE_RECURSIVE      = NO
