+++ documentation/documentation.dox
@@ -62,7 +62,7 @@ For more information, please refer to <https://unlicense.org> */
 //! @details Decorate, compose additional functionality atop filters while
 //! preserving the filter API.
 
-//! @namespace fcarouge::internal
+//! @namespace fcarouge::kalman_internal
 //!
 //! @brief Internal implementation details.
 //!
