+++ include/CMakeLists.txt
@@ -45,20 +45,20 @@ target_sources(
             "HEADERS"
             FILES
             "fcarouge/format.hpp"
-            "fcarouge/internal/factory.hpp"
-            "fcarouge/internal/function.hpp"
-            "fcarouge/internal/kalman.tpp"
-            "fcarouge/internal/type.hpp"
-            "fcarouge/internal/utility.hpp"
-            "fcarouge/internal/x_z_p_q_r_h_f.hpp"
-            "fcarouge/internal/x_z_p_q_r_hh_us_ps.hpp"
-            "fcarouge/internal/x_z_p_q_r.hpp"
-            "fcarouge/internal/x_z_p_qq_rr_f.hpp"
-            "fcarouge/internal/x_z_p_r_f.hpp"
-            "fcarouge/internal/x_z_p_r.hpp"
-            "fcarouge/internal/x_z_u_p_q_r_f_g_ps.hpp"
-            "fcarouge/internal/x_z_u_p_q_r_h_f_g_us_ps.hpp"
-            "fcarouge/internal/x_z_u_p_q_r.hpp"
+            "fcarouge/kalman_internal/factory.hpp"
+            "fcarouge/kalman_internal/function.hpp"
+            "fcarouge/kalman_internal/kalman.tpp"
+            "fcarouge/kalman_internal/type.hpp"
+            "fcarouge/kalman_internal/utility.hpp"
+            "fcarouge/kalman_internal/x_z_p_q_r_h_f.hpp"
+            "fcarouge/kalman_internal/x_z_p_q_r_hh_us_ps.hpp"
+            "fcarouge/kalman_internal/x_z_p_q_r.hpp"
+            "fcarouge/kalman_internal/x_z_p_qq_rr_f.hpp"
+            "fcarouge/kalman_internal/x_z_p_r_f.hpp"
+            "fcarouge/kalman_internal/x_z_p_r.hpp"
+            "fcarouge/kalman_internal/x_z_u_p_q_r_f_g_ps.hpp"
+            "fcarouge/kalman_internal/x_z_u_p_q_r_h_f_g_us_ps.hpp"
+            "fcarouge/kalman_internal/x_z_u_p_q_r.hpp"
             "fcarouge/kalman.hpp"
             "fcarouge/printer.hpp"
             "fcarouge/utility.hpp")
