+++ include/fcarouge/kalman_internal/x_z_p_q_r_h_f.hpp
@@ -36,36 +36,22 @@ OTHER DEALINGS IN THE SOFTWARE.
 
 For more information, please refer to <https://unlicense.org> */
 
-#ifndef FCAROUGE_INTERNAL_X_Z_U_P_Q_R_H_F_G_US_PS_HPP
-#define FCAROUGE_INTERNAL_X_Z_U_P_Q_R_H_F_G_US_PS_HPP
+#ifndef FCAROUGE_KALMAN_INTERNAL_X_Z_P_Q_R_H_F_HPP
+#define FCAROUGE_KALMAN_INTERNAL_X_Z_P_Q_R_H_F_HPP
 
 #include "fcarouge/utility.hpp"
-#include "function.hpp"
 
-#include <tuple>
-
-namespace fcarouge::internal {
-// Helper template to support multiple pack deduction.
-template <typename, typename, typename, typename, typename>
-struct x_z_u_p_q_r_h_f_g_us_ps final {};
-
-template <typename State, typename Output, typename Input,
-          typename... UpdateTypes, typename... PredictionTypes>
-struct x_z_u_p_q_r_h_f_g_us_ps<State, Output, Input, std::tuple<UpdateTypes...>,
-                               std::tuple<PredictionTypes...>> {
+namespace fcarouge::kalman_internal {
+template <typename State, typename Output> struct x_z_p_q_r_h_f {
   using state = State;
   using output = Output;
-  using input = Input;
   using estimate_uncertainty = ᴀʙᵀ<state, state>;
   using process_uncertainty = ᴀʙᵀ<state, state>;
   using output_uncertainty = ᴀʙᵀ<output, output>;
-  using state_transition = ᴀʙᵀ<state, state>;
+  using state_transition = evaluate<quotient<state, state>>;
   using output_model = ᴀʙᵀ<output, state>;
-  using input_control = ᴀʙᵀ<state, input>;
   using innovation = evaluate<difference<output, output>>;
   using innovation_uncertainty = output_uncertainty;
-  using update_types = std::tuple<UpdateTypes...>;
-  using prediction_types = std::tuple<PredictionTypes...>;
   using gain =
       evaluate<quotient<product<estimate_uncertainty, transpose<output_model>>,
                         innovation_uncertainty>>;
@@ -78,35 +64,25 @@ struct x_z_u_p_q_r_h_f_g_us_ps<State, Output, Input, std::tuple<UpdateTypes...>,
   output_uncertainty r{zero<output_uncertainty>};
   output_model h{one<output_model>};
   state_transition f{one<state_transition>};
-  input_control g{one<input_control>};
-  input u{zero<input>};
   gain k{one<gain>};
   innovation y{zero<innovation>};
   innovation_uncertainty s{one<innovation_uncertainty>};
   output z{zero<output>};
-  update_types update_arguments{};
-  prediction_types prediction_arguments{};
 
-  inline constexpr void update(const UpdateTypes &...update_pack,
-                               const auto &output_z, const auto &...outputs_z) {
-    update_arguments = {update_pack...};
+  inline constexpr void update(const auto &output_z, const auto &...outputs_z) {
     z = output{output_z, outputs_z...};
-    s = h * p * t(h) + r;
+    s = innovation_uncertainty{h * p * t(h) + r};
     k = p * t(h) / s;
     y = z - h * x;
-    x = x + k * y;
-    p = (i - k * h) * p * t(i - k * h) + k * r * t(k);
+    x = state{x + k * y};
+    p = estimate_uncertainty{(i - k * h) * p * t(i - k * h) + k * r * t(k)};
   }
 
-  //! @todo Add convertible requirements on input and output packs?
-  inline constexpr void predict(const PredictionTypes &...prediction_pack,
-                                const auto &input_u, const auto &...inputs_u) {
-    prediction_arguments = {prediction_pack...};
-    u = input{input_u, inputs_u...};
-    x = f * x + g * u;
-    p = f * p * t(f) + q;
+  inline constexpr void predict() {
+    x = f * x;
+    p = estimate_uncertainty{f * p * t(f) + q};
   }
 };
-} // namespace fcarouge::internal
+} // namespace fcarouge::kalman_internal
 
-#endif // FCAROUGE_INTERNAL_X_Z_U_P_Q_R_H_F_G_US_PS_HPP
+#endif // FCAROUGE_KALMAN_INTERNAL_X_Z_P_Q_R_H_F_HPP
