+++ include/fcarouge/kalman_internal/x_z_u_p_q_r_h_f_g_us_ps.hpp
@@ -36,53 +36,77 @@ OTHER DEALINGS IN THE SOFTWARE.
 
 For more information, please refer to <https://unlicense.org> */
 
-#ifndef FCAROUGE_INTERNAL_X_Z_U_P_Q_R_HPP
-#define FCAROUGE_INTERNAL_X_Z_U_P_Q_R_HPP
+#ifndef FCAROUGE_KALMAN_INTERNAL_X_Z_U_P_Q_R_H_F_G_US_PS_HPP
+#define FCAROUGE_KALMAN_INTERNAL_X_Z_U_P_Q_R_H_F_G_US_PS_HPP
 
 #include "fcarouge/utility.hpp"
 #include "function.hpp"
 
 #include <tuple>
 
-namespace fcarouge::internal {
-template <typename Type> struct x_z_u_p_q_r {
-  using state = Type;
-  using output = Type;
-  using input = Type;
+namespace fcarouge::kalman_internal {
+// Helper template to support multiple pack deduction.
+template <typename, typename, typename, typename, typename>
+struct x_z_u_p_q_r_h_f_g_us_ps final {};
+
+template <typename State, typename Output, typename Input,
+          typename... UpdateTypes, typename... PredictionTypes>
+struct x_z_u_p_q_r_h_f_g_us_ps<State, Output, Input, std::tuple<UpdateTypes...>,
+                               std::tuple<PredictionTypes...>> {
+  using state = State;
+  using output = Output;
+  using input = Input;
   using estimate_uncertainty = ᴀʙᵀ<state, state>;
   using process_uncertainty = ᴀʙᵀ<state, state>;
   using output_uncertainty = ᴀʙᵀ<output, output>;
+  using state_transition = ᴀʙᵀ<state, state>;
+  using output_model = ᴀʙᵀ<output, state>;
+  using input_control = ᴀʙᵀ<state, input>;
   using innovation = evaluate<difference<output, output>>;
   using innovation_uncertainty = output_uncertainty;
-  using gain = evaluate<quotient<estimate_uncertainty, innovation_uncertainty>>;
+  using update_types = std::tuple<UpdateTypes...>;
+  using prediction_types = std::tuple<PredictionTypes...>;
+  using gain =
+      evaluate<quotient<product<estimate_uncertainty, transpose<output_model>>,
+                        innovation_uncertainty>>;
 
-  static inline const auto i{one<gain>};
+  static inline const auto i{one<ᴀʙᵀ<state, state>>};
 
   state x{zero<state>};
   estimate_uncertainty p{one<estimate_uncertainty>};
   process_uncertainty q{zero<process_uncertainty>};
   output_uncertainty r{zero<output_uncertainty>};
+  output_model h{one<output_model>};
+  state_transition f{one<state_transition>};
+  input_control g{one<input_control>};
   input u{zero<input>};
   gain k{one<gain>};
   innovation y{zero<innovation>};
   innovation_uncertainty s{one<innovation_uncertainty>};
   output z{zero<output>};
+  update_types update_arguments{};
+  prediction_types prediction_arguments{};
 
-  inline constexpr void update(const auto &output_z, const auto &...outputs_z) {
+  inline constexpr void update(const UpdateTypes &...update_pack,
+                               const auto &output_z, const auto &...outputs_z) {
+    update_arguments = {update_pack...};
     z = output{output_z, outputs_z...};
-    s = p + r;
-    k = p / s;
-    y = z - x;
+    s = h * p * t(h) + r;
+    k = p * t(h) / s;
+    y = z - h * x;
     x = x + k * y;
-    p = (i - k) * p * t(i - k) + k * r * t(k);
+    p = (i - k * h) * p * t(i - k * h) + k * r * t(k);
   }
 
-  inline constexpr void predict(const auto &input_u, const auto &...inputs_u) {
+  //! @todo Add convertible requirements on input and output packs?
+  inline constexpr void predict(const PredictionTypes &...prediction_pack,
+                                const auto &input_u, const auto &...inputs_u) {
+    prediction_arguments = {prediction_pack...};
     u = input{input_u, inputs_u...};
-    x = u;
-    p = p + q;
+    x = f * x + g * u;
+    p = f * p * t(f) + q;
   }
 };
-} // namespace fcarouge::internal
+} // namespace fcarouge::kalman_internal
 
-#endif // FCAROUGE_INTERNAL_X_Z_U_P_Q_R_HPP
+#endif // FCAROUGE_KALMAN_INTERNAL_X_Z_U_P_Q_R_H_F_G_US_PS_HPP
